% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enadid_2006.R
\docType{data}
\name{enadid_2006}
\alias{enadid_2006}
\title{National Survey of Demographic Dynamics 2006}
\format{A data frame with 91000 rows and 13 variables


* ID_1: Indicate the woman's identification.

* FACTOR: Sampling weights.

* UPM: Cluster ids.

* ESTRATO: Strata.

* FEC_MUJ_M: The woman’s month of birth.

* FEC_MUJ_A: The woman’s year of birth.

* EDAD_M: The woman's age at the time of the interview (mother or not mother).

* ENTREV_M: The month in which the mother were interviewed.

* NUM_HIJ: The woman’s number of children.

* FEC_HIJ_M: The child’s month of birth.

* FEC_HIJ_A: The child’s year of birth.

* CONT: Indicate 0 if the woman doesn't have a child or 1 if the woman has a child.

* MUJER: Indicate TRUE if the woman isn't duplicate or FALSE if the woman is duplicate.}
\source{
<http://www.conapo.gob.mx/es/CONAPO/ENADID_2006>
}
\usage{
enadid_2006
}
\description{
A dataset containing  attributes of the women and attributes of the children.
}
\keyword{datasets}
