% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enadid_2009.R
\docType{data}
\name{enadid_2009}
\alias{enadid_2009}
\title{National Survey of Demographic Dynamics 2009}
\format{A data frame with 220259 rows and 13 variables


* ID_1: Indicate the woman's identification.

* FACTOR: Sampling weights.

* UPM: Cluster ids.

* ESTRATO: Strata.

* FEC_MUJ_M: The woman’s month of birth.

* FEC_MUJ_A: The woman’s year of birth.

* EDAD_M: The woman's age at the time of the interview (mother or not mother).

* ENTREV_M: The month in which the mother were interviewed.

* NUM_HIJ: The woman’s number of children.

* FEC_HIJ_M: The child’s month of birth.

* FEC_HIJ_A: The child’s year of birth.

* CONT: Indicate 0 if the woman doesn't have a child or 1 if the woman has a child.

* MUJER: Indicate TRUE if the woman isn't duplicate or FALSE if the woman is duplicate.}
\source{
<https://www.inegi.org.mx/programas/enadid/2009/default.html#Microdatos>
}
\usage{
enadid_2009
}
\description{
A dataset containing  attributes of the women and attributes of the children.
}
\keyword{datasets}
