% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enadid.R
\docType{data}
\name{enadid_2014}
\alias{enadid_2014}
\title{National Survey of Demographic Dynamics 2014}
\format{A data frame with 214335 rows and 13 variables}
\source{
<http://www.beta.inegi.org.mx/proyectos/enchogares/especiales/enadid/2014/>
}
\usage{
enadid_2014
}
\description{
A dataset containing  attributes of the women and attributes of the children.
The variables are as follows:
}
\details{
* `ID_1`:
* `FACTOR`: j
* `UPM`:
* `ESTRATO`:
* `ENTIDAD`:
* `FEC_MUJ_M`:
* `FEC_MUJ_A`:
* `ENTREV_M`:
* `NUM_HIJ`:
* `FEC_HIJ_M`:
* `FEC_HIJ_A`:
* `CONT`:
* `MUJER`:
}
\keyword{datasets}
