% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frts_15yrs.R
\name{frts_15yrs}
\alias{frts_15yrs}
\title{Apply fertility rates methodology for last fifteen-years to annual periods}
\usage{
frts_15yrs(y.intvw, m.wmn, y.wmn, m.child, y.child, children, child.dummy,
  wmn.dummy, id.wmn, weights, data, ...)
}
\arguments{
\item{y.intvw}{A numeric which indicate the year in which the mother were interviewed.}

\item{m.wmn}{A vector specifying the woman’s month of birth (mother or not mother).}

\item{y.wmn}{A vector specifying the woman’s year of birth (mother or not mother).}

\item{m.child}{A vector specifying the child’s month of birth (if the woman doesn't have child, NA).}

\item{y.child}{A vector specifying the child’s year of birth (if the woman doesn't have children, NA).}

\item{children}{A vector which indicate the number of children.}

\item{child.dummy}{A vector which indicate 0 if the woman doesn't have a child or 1 if the woman has a child.}

\item{wmn.dummy}{A vector which indicate TRUE if the woman isn't duplicate or FALSE if the woman is duplicate.}

\item{id.wmn}{A vector wich indicate the woman's identification.}

\item{weights}{A vetor pecifying sampling weights as an alternative to prob (1/weights).}

\item{data}{A data frame containing the above variables.}

\item{ids}{A vector specifying cluster ids from largest level to smallest level.}

\item{strata}{A vector specifying strata.}
}
\value{
An object of class frts_15yrs containing four arrays and a numeric.
}
\description{
Applies a given fertility rates methodology rates
case-by-case to a data set and returns a list with the estimates by year.
}
\examples{

## Information from ENADID 2014, INEGI

mg4 <- frts_15yrs(y.intvw=2014, m.wmn=FEC_MUJ_M, y.wmn=FEC_MUJ_A, m.child=FEC_HIJ_M,
y.child=FEC_HIJ_A, children=NUM_HIJ, child.dummy=CONT,wmn.dummy=MUJER, id.wmn=ID_1,
data = enadid_2014, weights = FACTOR)

summary(mg4)

}
