% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frts_intvw.R
\name{frts_intvw}
\alias{frts_intvw}
\title{Apply fertility rates methodology for year of interview}
\usage{
frts_intvw(m.intvw, y.intvw, m.wmn, y.wmn, age.wmn, m.child, y.child,
  child.dummy, wmn.dummy, id.wmn, ids, strata, weights, data)
}
\arguments{
\item{m.intvw}{A numeric or a vector which indicates the month in which the woman were interviewed.}

\item{y.intvw}{A numeric which indicate the year in which the woman were interviewed.}

\item{m.wmn}{A vector specifying the woman’s month of birth (mother or not mother).}

\item{y.wmn}{A vector specifying the woman’s year of birth (mother or not mother).}

\item{age.wmn}{A vector specifying the woman's age at the time of the interview (mother or not mother).}

\item{m.child}{A vector specifying the child’s month of birth (if the woman doesn't have child, NA).}

\item{y.child}{A vector specifying the child’s year of birth (if the woman doesn't have children, NA).}

\item{child.dummy}{A vector which indicate 0 if the woman doesn't have a child or 1 if the woman has a child.}

\item{wmn.dummy}{A vector which indicate TRUE if the woman isn't duplicate or FALSE if the woman is duplicate.}

\item{id.wmn}{A vector wich indicate the woman's identification.}

\item{ids}{A vector specifying cluster ids from largest level to smallest level.}

\item{strata}{A vector specifying strata.}

\item{weights}{A vetor pecifying sampling weights as an alternative to prob (1/weights).}

\item{data}{A data frame containing the above variables.}
}
\value{
An object of class frts_intvw containing a data.frame and a list.
}
\description{
Applies a given fertility rates methodology rates
case-by-case to a data set and returns another data set
with the estimates by observation.
}
\examples{

## Information from ENADID 2014, INEGI

mg1 <- frts_intvw(m.intvw=ENTREV_M, y.intvw=2014, m.wmn=FEC_MUJ_M, y.wmn=FEC_MUJ_A, age.wmn=EDADD, m.child=FEC_HIJ_M,
y.child=FEC_HIJ_A, child.dummy=CONT,wmn.dummy=MUJER, id.wmn=ID_1, ids=UPM, strata=ESTRATO,
weights = FACTOR, data = enadid_2014)

summary(mg1, level = 0.9)

}
