% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frts_yrly.R
\name{frts_yrly}
\alias{frts_yrly}
\title{Apply fertility rates methodology for last complete years up to five years}
\usage{
frts_yrly(m.intvw, y.intvw, y.ref, m.wmn, y.wmn, age.wmn, m.child, y.child,
  wmn.dummy, id.wmn, ids, strata, weights, data)
}
\arguments{
\item{m.intvw}{A numeric or a vector which indicates the month in which the woman were interviewed.}

\item{y.intvw}{A numeric which indicate the year in which the woman were interviewed.}

\item{y.ref}{A numeric which indicate the reference year. It must be between last five years above interview.}

\item{m.wmn}{A vector specifying the woman’s month of birth (mother or not mother).}

\item{y.wmn}{A vector specifying the woman’s year of birth (mother or not mother).}

\item{age.wmn}{A vector specifying the woman's age at the time of the interview (mother or not mother).}

\item{m.child}{A vector specifying the child’s month of birth (if the woman doesn't have child, NA).}

\item{y.child}{A vector specifying the child’s year of birth (if the woman doesn't have children, NA).}

\item{wmn.dummy}{A vector which indicate TRUE if the woman isn't duplicate or FALSE if the woman is duplicate.}

\item{id.wmn}{A vector wich indicate the woman's identification.}

\item{ids}{A vector specifying cluster ids from largest level to smallest level, ~0 or ~1 is a formula for no clusters.}

\item{strata}{A vector specifying strata.}

\item{weights}{A vetor pecifying sampling weights as an alternative to prob (1/weights).}

\item{data}{A data frame containing the above variables.}
}
\value{
An object of class frts_yrly containing a data.frame, a list and a numeric.
}
\description{
Applies a given fertility rates methodology rates
case-by-case to a data set and returns another data set
with the estimates by observation.
}
\examples{

## Information from ENADID 2014, INEGI

mg2 <- frts_yrly(m.intvw=ENTREV_M, y.intvw=2014, y.ref=2012, m.wmn=FEC_MUJ_M,
 y.wmn=FEC_MUJ_A, age.wmn=EDAD_M, m.child=FEC_HIJ_M, y.child=FEC_HIJ_A,
 wmn.dummy=MUJER, id.wmn=ID_1, ids=UPM, strata=ESTRATO, weights = FACTOR,
 data = enadid_2014)

summary(mg2, level = 0.9)

}
