% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.frts_15yrs}
\alias{summary.frts_15yrs}
\title{Summarize the data of last fifteen-years to annual periods in fertility rates}
\usage{
\method{summary}{frts_15yrs}(x)
}
\arguments{
\item{x}{An object of class "frts_3yrs".}
}
\value{
An object of class list. The elements of this are "as_fr_s", "t_fr" and "as_fr_g".
}
\description{
Summarize the data of last fifteen-years to annual periods in age-specific fertility rates
by single year of age and grouped year of age, and total fertility rate.
}
