% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.frts_3yrs}
\alias{summary.frts_3yrs}
\title{Summarize the data of last five years to three-year periods in fertility rates}
\usage{
\method{summary}{frts_3yrs}(x, level)
}
\arguments{
\item{x}{An object of class "frts_3yrs".}

\item{level}{The confidence level required to confidence intervals for fertility rates.}
}
\value{
An object of class list. The elements of this are "as_fr_s", "t_fr" and "as_fr_g".
}
\description{
Summarize the data of last five years to three-year periods in age-specific fertility rates
by single year of age and grouped year of age, and total fertility rate.
}
\examples{

## Information from ENADID 2014, INEGI

mg3 <- frts_3yrs(y.first=2013, y.second=2012, y.third=2011, m.wmn=FEC_MUJ_M, y.wmn=FEC_MUJ_A,
m.child=FEC_HIJ_M, y.child=FEC_HIJ_A, children=NUM_HIJ, child.dummy=CONT,wmn.dummy=MUJER,
id.wmn=ID_1, ids=UPM, strata=ESTRATO, data = enadid_2014, weights = FACTOR)

summary(mg3, level = 0.9)

}
