% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.frts_yrly}
\alias{summary.frts_yrly}
\title{Summarize the data of last complete years up to five years in fertility rates}
\usage{
\method{summary}{frts_yrly}(object, level, ...)
}
\arguments{
\item{object}{An object of class "frts_yrly".}

\item{level}{The confidence level required to confidence intervals for fertility rates.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
An object of class list. The elements of this are "as_fr_s", "t_fr" and "as_fr_g".
}
\description{
Summarize the data of last complete years up to five years in age-specific fertility rates
by single year of age and grouped year of age, and total fertility rate.
}
\examples{

## Information from ENADID 2014, INEGI

mg2 <- frts_yrly(m.intvw=ENTREV_M, y.intvw=2014, y.ref=2012, m.wmn=FEC_MUJ_M, y.wmn=FEC_MUJ_A, age.wmn=EDAD_M, m.child=FEC_HIJ_M,
 y.child=FEC_HIJ_A, wmn.dummy=MUJER, id.wmn=ID_1,
 ids=UPM, strata=ESTRATO, weights = FACTOR, data = enadid_2014)

summary(mg2, level = 0.9)
}
